refactor: remove legacy code, rename web_service_v3 to web_service
- Remove action_provider/, common/, lambda/, examples/, docs/ap/
- Rename web_service_v3/ to web_service/ as the canonical service
- Merge common/ utilities into web_service/utils.py
- Move test files from web_service_v3/ to tests/ with *_test.py naming
- Remove old v2 web_service code and action_provider tests
- Update Dockerfile, CI workflows, docs, and pyproject.toml accordingly
