+++ .gitignore
@@ -128,7 +128,7 @@ dmypy.json
 # Pyre type checker
 .pyre/
 
-# Diaspora Web Service and Action Provider
+# Diaspora Web Service
 secrets.sh
 nohup.out
 *.pyc
