--- action_provider/__init__.py
@@ -1,3 +1,0 @@
-"""Diaspora Action Provider."""
-
-from __future__ import annotations
