--- action_provider/py.typed
@@ -1,1 +1,0 @@
-# PEP 561 marker file to indicate this package is typed
