--- action_provider/schema.json
@@ -1,115 +1,0 @@
-{
-    "title": "Diaspora Event Fabric Messaging Schema",
-    "type": "object",
-    "properties": {
-        "action": {
-            "description": "'produce' to publish events, 'consume' to retrieve events.",
-            "type": "string",
-            "enum": [
-                "produce",
-                "consume"
-            ]
-        },
-        "topic": {
-            "description": "Topic to publish or retrieve events.",
-            "type": "string"
-        },
-        "key": {
-            "type": "string",
-            "description": "Optional key paired with 'value' for 'produce'."
-        },
-        "value": {
-            "type": "object",
-            "description": "Single JSON event. Use either 'value' or 'msgs'."
-        },
-        "keys": {
-            "oneOf": [
-                {
-                    "type": "string",
-                    "description": "Optional key paired with 'msgs' for 'produce'."
-                },
-                {
-                    "type": "array",
-                    "description": "Optional list of keys paired with 'msgs' for 'produce'.",
-                    "items": {
-                        "type": "string"
-                    }
-                }
-            ]
-        },
-        "msgs": {
-            "type": "array",
-            "description": "List of JSON events. Use either 'value' or 'msgs'.",
-            "items": {
-                "type": "object"
-            }
-        },
-        "ts": {
-            "description": "Epoch timestamp to retrieve messages from (optional for 'consume').",
-            "type": "integer",
-            "minimum": 0
-        },
-        "group_id": {
-            "description": "Group ID for the message (optional for 'consume').",
-            "type": "string"
-        },
-        "filters": {
-            "type": "array",
-            "description": "Filters for retrieving events (optional for 'consume').",
-            "items": {
-                "type": "object"
-            }
-        },
-        "servers": {
-            "description": "List of diaspora servers (dev use).",
-            "type": "string"
-        }
-    },
-    "additionalProperties": false,
-    "required": [
-        "action",
-        "topic"
-    ],
-    "dependencies": {
-        "action": {
-            "oneOf": [
-                {
-                    "properties": {
-                        "action": {
-                            "const": "produce"
-                        },
-                        "msgs": {
-                            "type": "array",
-                            "minItems": 1
-                        },
-                        "value": {
-                            "type": "object"
-                        }
-                    },
-                    "oneOf": [
-                        {
-                            "required": [
-                                "msgs"
-                            ]
-                        },
-                        {
-                            "required": [
-                                "value"
-                            ]
-                        }
-                    ]
-                },
-                {
-                    "properties": {
-                        "action": {
-                            "const": "consume"
-                        },
-                        "ts": {
-                            "type": "integer"
-                        }
-                    }
-                }
-            ]
-        }
-    }
-}
