--- common/__init__.py
@@ -1,7 +1,0 @@
-"""Diaspora Action Provider and Web Service common functions."""
-
-from __future__ import annotations
-
-import importlib.metadata as importlib_metadata
-
-__version__ = importlib_metadata.version('diaspora_service')
