+++ docs/2-testing.md
@@ -2,30 +2,25 @@
 
 ## Testing in Virtual Environment
 
-### Action Provider
-
-To test the Action Provider in a virtual environment, run:
-
-```bash
-pytest tests/action_provider_test.py
-
-coverage erase
-coverage run -m pytest
-coverage report
-```
-
-
 ### Web Service
 
 To test the Web Service in a virtual environment, run:
 
 ```bash
-pytest tests/web_service_test.py
+pytest tests/ -vv -s
 ```
 
-### Run `tox` for both services
+With coverage:
 
 ```bash
-source secrets.sh
+coverage erase
+coverage run -m pytest tests/
+coverage report
+```
+
+### Run `tox`
+
+```bash
+source secrets2.sh  # pragma: allowlist secret
 tox
 ```
