+++ docs/4-monitoring.md
@@ -2,6 +2,4 @@
 
 1. Check the service live status at [Service Status](https://haochenpan.github.io/diaspora-uptime-monitor/) dashboard.
 
-2. The [Lambda function](https://github.com/haochenpan/diaspora-service/blob/main/lambda/lambda_function.py) reboots the containers as needed.
-
-3. Use [Diaspora Admin Console](kafka-console/intro.md)
+2. Use [Diaspora Admin Console](kafka-console/intro.md)
