+++ docs/acknowledgments.md
@@ -1,5 +1,3 @@
 # Acknowledgments
 
 This project's layout and workflows are primarily based on [Greg's Python template](https://github.com/gpauloski/python-template).
-
-The implementation of the action provider follows Will's [ALCF Action Provider](https://github.com/globus-labs/alcf-status-action-provider).
