--- docs/ap/api.md
@@ -1,4 +1,0 @@
-## [live schema](https://diaspora-action-provider.ml22sevubfnks.us-east-1.cs.amazonlightsail.com/)
-
-<!-- ## - action_provider.main.action_run -->
-## ::: action_provider.main.action_run
