--- docs/ap/examples.md
@@ -1,2 +1,0 @@
-- [Send requests to an AP Endpoint](https://github.com/haochenpan/diaspora-service/blob/main/examples/demo_to_host.ipynb)
-- [Use Diaspora AP in Globus Flows](https://github.com/haochenpan/diaspora-service/blob/main/examples/demo_to_flows.ipynb)
