+++ web_service/__init__.py
@@ -1,3 +1,7 @@
 """Diaspora Web Service."""
 
 from __future__ import annotations
+
+import importlib.metadata as importlib_metadata
+
+__version__ = importlib_metadata.version('diaspora_service')
