--- web_service_v3/__init__.py
@@ -1,3 +1,0 @@
-"""Diaspora Web Service."""
-
-from __future__ import annotations
